NH047 - Reimplement logic as low-scale BigDecimals

+++ nhood-engine-core/src/main/java/com/h8/nh/nhoodengine/core/matrix/DataMatrixResource.java
@@ -1,6 +1,8 @@
 package com.h8.nh.nhoodengine.core.matrix;
 
+import java.math.BigDecimal;
+
 public interface DataMatrixResource {
 
-    double[] getKey();
+    BigDecimal[] getKey();
 }
